Move set foreign key into separate folder
